Remove input text and close keyboard when submitted. Refactor.
