+++ components/AddRow.js
@@ -1,11 +1,23 @@
 import { useState } from "react";
-import { View, TextInput, StyleSheet } from "react-native";
+import { View, TextInput, StyleSheet, Keyboard } from "react-native";
 
 import { CountableButton } from "./CountableButton";
 import { CommonStyles } from "../styles/CommonStyles";
 
-export const AddRow = ({ handleSubmit }) => {
+export const AddRow = ({ addNewCountable, countables }) => {
   const [name, setName] = useState("");
+
+  const isValidInput = (name) => {
+    return !countables.some((countable) => countable.name === name);
+  };
+
+  const handleSubmit = (name) => {
+    if (isValidInput(name)) {
+      Keyboard.dismiss();
+      addNewCountable(name);
+      setName("");
+    }
+  };
 
   return (
     <View style={CommonStyles.row}>
@@ -13,6 +25,7 @@ export const AddRow = ({ handleSubmit }) => {
         style={styles.textInput}
         placeholder="Enter name"
         onChangeText={setName}
+        value={name}
       />
       <CountableButton
         label="Add"
